Add files via upload
The "Gold Price Prediction using Machine Learning model" project is a practical example of how machine learning can be applied to predict the price of gold. The project involves using historical data on gold prices and other relevant factors to build a predictive model using the Random Forest algorithm in Python.

The goal of the project is to develop a model that can accurately predict the price of gold in the future based on a set of input features. The project starts with data collection, which involves gathering historical data on gold prices, economic indicators, and other factors that may affect the price of gold.